+++ src/tcpedit/plugins/dlt_linuxsll/linuxsll.h
@@ -33,14 +33,15 @@ int dlt_linuxsll_register(tcpeditdlt_t *ctx);
 int dlt_linuxsll_init(tcpeditdlt_t *ctx);
 int dlt_linuxsll_cleanup(tcpeditdlt_t *ctx);
 int dlt_linuxsll_parse_opts(tcpeditdlt_t *ctx);
-int dlt_linuxsll_decode(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_linuxsll_encode(tcpeditdlt_t *ctx, u_char *packet, int pktlen, tcpr_dir_t dir);
-int dlt_linuxsll_proto(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_linuxsll_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen);
-u_char *dlt_linuxsll_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen, u_char *l3data);
+int dlt_linuxsll_decode(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_linuxsll_encode(tcpeditdlt_t *ctx, u_char *packet, size_t pktlen, tcpr_dir_t dir);
+int dlt_linuxsll_proto(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_linuxsll_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen);
+u_char *dlt_linuxsll_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen, u_char *l3data);
 tcpeditdlt_l2addr_type_t dlt_linuxsll_l2addr_type(void);
-int dlt_linuxsll_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_linuxsll_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac, const u_char *packet, const int pktlen);
+int dlt_linuxsll_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_linuxsll_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac,
+        const u_char *packet, const size_t pktlen);
 
 #ifdef __cplusplus
 }
