+++ src/tcpedit/plugins/dlt_null/null.h
@@ -31,14 +31,15 @@ int dlt_null_register(tcpeditdlt_t *ctx);
 int dlt_null_init(tcpeditdlt_t *ctx);
 int dlt_null_cleanup(tcpeditdlt_t *ctx);
 int dlt_null_parse_opts(tcpeditdlt_t *ctx);
-int dlt_null_decode(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_null_encode(tcpeditdlt_t *ctx, u_char *packet, int pktlen, tcpr_dir_t dir);
-int dlt_null_proto(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_null_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen);
-u_char *dlt_null_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen, u_char *l3data);
+int dlt_null_decode(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_null_encode(tcpeditdlt_t *ctx, u_char *packet, size_t pktlen, tcpr_dir_t dir);
+int dlt_null_proto(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_null_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen);
+u_char *dlt_null_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen, u_char *l3data);
 tcpeditdlt_l2addr_type_t dlt_null_l2addr_type(void);
-int dlt_null_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_null_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac, const u_char *packet, const int pktlen);
+int dlt_null_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_null_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac,
+        const u_char *packet, const size_t pktlen);
 
 #endif
