+++ src/tcpedit/plugins/dlt_pppserial/pppserial.h
@@ -28,14 +28,16 @@ int dlt_pppserial_init(tcpeditdlt_t *ctx);
 int dlt_pppserial_post_init(tcpeditdlt_t *ctx);
 int dlt_pppserial_cleanup(tcpeditdlt_t *ctx);
 int dlt_pppserial_parse_opts(tcpeditdlt_t *ctx);
-int dlt_pppserial_decode(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_pppserial_encode(tcpeditdlt_t *ctx, u_char *packet, int pktlen, tcpr_dir_t dir);
-int dlt_pppserial_proto(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_pppserial_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen);
-u_char *dlt_pppserial_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen, u_char *l3data);
+int dlt_pppserial_decode(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_pppserial_encode(tcpeditdlt_t *ctx, u_char *packet, size_t pktlen, tcpr_dir_t dir);
+int dlt_pppserial_proto(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_pppserial_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen);
+u_char *dlt_pppserial_merge_layer3(tcpeditdlt_t *ctx, u_char *packet,
+        const size_t pktlen, u_char *l3data);
 tcpeditdlt_l2addr_type_t dlt_pppserial_l2addr_type(void);
-int dlt_pppserial_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_pppserial_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac, const u_char *packet, const int pktlen);
+int dlt_pppserial_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_pppserial_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac,
+        const u_char *packet, const size_t pktlen);
 
 #endif
