+++ src/tcpedit/plugins/dlt_radiotap/radiotap.h
@@ -28,15 +28,17 @@ int dlt_radiotap_register(tcpeditdlt_t *ctx);
 int dlt_radiotap_init(tcpeditdlt_t *ctx);
 int dlt_radiotap_cleanup(tcpeditdlt_t *ctx);
 int dlt_radiotap_parse_opts(tcpeditdlt_t *ctx);
-int dlt_radiotap_decode(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_radiotap_encode(tcpeditdlt_t *ctx, u_char *packet, int pktlen, tcpr_dir_t dir);
-int dlt_radiotap_proto(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_radiotap_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen);
-u_char *dlt_radiotap_merge_layer3(tcpeditdlt_t *ctx, u_char *packet, const int pktlen, u_char *l3data);
+int dlt_radiotap_decode(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_radiotap_encode(tcpeditdlt_t *ctx, u_char *packet, size_t pktlen, tcpr_dir_t dir);
+int dlt_radiotap_proto(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_radiotap_get_layer3(tcpeditdlt_t *ctx, u_char *packet, const size_t pktlen);
+u_char *dlt_radiotap_merge_layer3(tcpeditdlt_t *ctx, u_char *packet,
+        const size_t pktlen, u_char *l3data);
 tcpeditdlt_l2addr_type_t dlt_radiotap_l2addr_type(void);
-int dlt_radiotap_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-int dlt_radiotap_80211_l2len(tcpeditdlt_t *ctx, const u_char *packet, const int pktlen);
-u_char *dlt_radiotap_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac, const u_char *packet, const int pktlen);
+int dlt_radiotap_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+int dlt_radiotap_80211_l2len(tcpeditdlt_t *ctx, const u_char *packet, const size_t pktlen);
+u_char *dlt_radiotap_get_mac(tcpeditdlt_t *ctx, tcpeditdlt_mac_type_t mac,
+        const u_char *packet, const size_t pktlen);
 
 /*
  * FIXME: structure to hold any data parsed from the packet by the decoder.
